api/requirements.txt
fastapi
uvicorn[standard]
tensorflow>=2.11
pillow
numpy
pydantic
